REFACTOR: Consolidate functions for unpacking `RpcResult`s.
