+++ packages/frontend/src/api/document_rpc.test.ts
@@ -5,8 +5,8 @@ import { deleteUser, getAuth, signOut } from "firebase/auth";
 import * as uuid from "uuid";
 import { assert, afterAll, describe, test } from "vitest";
 
-import { initTestUserAuth, unwrap, unwrapErr } from "../util/test_util.ts";
-import { createRpcClient } from "./rpc.ts";
+import { initTestUserAuth } from "../util/test_util.ts";
+import { createRpcClient, unwrap, unwrapErr } from "./rpc.ts";
 
 const serverUrl = import.meta.env.VITE_SERVER_URL;
 const repoUrl = import.meta.env.VITE_AUTOMERGE_REPO_URL;
